Added several changes for more functionality
1: Allowed for robot to measure the distance of a square as it is
moving across it. This distance is measured in encoder ticks and is
saved in a ChessBot data member (squareDistance). This data member will
be used to move distances proportional to the chessboard square's
length.

2. Simplified the AlignToEdge() function, which previously contained a
routine for aligning to a square edge and reversing half the square's
length. Now it only contains the routine to aligning to an edge.

3. Added function MoveDistance() to move certain distances in encoder
counts. This compensates for the simplification of the AlignToEdge()
function.